+++ ledger/complete/common/triequeue_test.go
@@ -1,4 +1,4 @@
-package wal
+package common
 
 import (
 	"math/rand"
